+++ app/pages/index.pug
@@ -21,4 +21,4 @@ html(lang="ru")
             include ./script-templates/loadphoto
             
         script(src="https://cdn.jsdelivr.net/npm/handlebars@latest/dist/handlebars.js")
-        script(defer src="main.min.js")
+        script(defer src="./scripts/main.js" type="module")
