+++ app/pages/script-templates/loadphoto.pug
@@ -1,7 +1,7 @@
 #loadhoto.loadphoto.hidden
     .container
         .loadphoto__wrapper
-            .loadphoto__avatar
+            label.loadphoto__avatar
                 #dragContainer.loadphoto__image-container(draggable="true")
                     input(type="file").loadphoto__input
                     #loadPhotoImage
